Implement if statements
If statements consist of the `if` keyword followed by a body--an
arbitrary number of expressions wrapped in curly brackets. If statements
pop an element which must be a boolean off the stack and conditionally
execute the expressions within their body if the value popped off the
stack is true.

Example:
```c-like
true if {
    42 print
}
```

If statements should eventually support following else statements.